# openapi.yaml
openapi: 3.1.0
info:
  title: etokendb API
  version: 0.1.0
  summary: Public API for token metrics, trade history, paid reviews, candles, and server-side access analytics
  description: |
    etokendb exposes a JSON API backed by SQLite.

    Response envelopes:

    - Success: `{ "ok": true, "data": ... }`
    - Error: `{ "ok": false, "error": { "code": "...", "message": "..." } }`

    General rules:

    - Endpoints are GET only except the documented review invoice POST routes
    - Unsupported methods return `405 METHOD_NOT_ALLOWED`
    - Counts are numbers
    - Large satoshi and price values are strings to preserve precision
    - Analytics time buckets use Unix milliseconds
    - Trade block timestamps and candle bucket timestamps use Unix seconds
servers:
  - url: http://127.0.0.1:8787
    description: Local default server
  - url: https://etokendb.alitayin.com
    description: Production example
...
paths:
...
  /api/tokens:
    get:
      tags: [Tokens]
      summary: List token summaries
      operationId: listTokens
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: sort
          in: query
          required: false
          description: Sort field. Effective default is `recent144VolumeSats`.
          schema:
            $ref: "#/components/schemas/TokenSortField"
        - name: order
          in: query
          required: false
          description: Sort order. Effective default is `desc`.
          schema:
            $ref: "#/components/schemas/SortOrder"
        - name: readyOnly
          in: query
          required: false
          description: |
            Filter to ready tokens only. If omitted, the current implementation behaves as if `true`.
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Paginated token summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTokenSummaryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /api/tokens/{tokenId}:
    get:
      tags: [Tokens]
      summary: Get a single token
      operationId: getToken
      parameters:
        - $ref: "#/components/parameters/TokenIdPathParam"
      responses:
        "200":
          description: Token detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenDetailResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /api/tokens/{tokenId}/trades:
    get:
      tags: [Tokens, Trades]
      summary: List trade history for one token
      operationId: listTokenTrades
      parameters:
        - $ref: "#/components/parameters/TokenIdPathParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated token trade history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTradeHistoryResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
...
  /api/tokens/{tokenId}/reviews:
    get:
      tags: [Tokens, Reviews]
      summary: List paid reviews for one token
      operationId: listTokenReviews
      parameters:
        - $ref: "#/components/parameters/TokenIdPathParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated paid reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTokenReviewsResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /api/tokens/{tokenId}/reviews/summary:
    get:
      tags: [Tokens, Reviews]
      summary: Get paid review score summary for one token
      operationId: getTokenReviewSummary
      parameters:
        - $ref: "#/components/parameters/TokenIdPathParam"
      responses:
        "200":
          description: Paid review score summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenReviewSummaryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /api/tokens/{tokenId}/reviews/invoices:
    post:
      tags: [Tokens, Reviews]
      summary: Create a paid review invoice
      operationId: createReviewInvoice
      parameters:
        - $ref: "#/components/parameters/TokenIdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewInvoiceRequest"
      responses:
        "201":
          description: Review invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewInvoiceResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
  /api/review-invoices/{invoiceId}:
    get:
      tags: [Reviews]
      summary: Get one review invoice
      operationId: getReviewInvoice
      parameters:
        - $ref: "#/components/parameters/InvoiceIdPathParam"
      responses:
        "200":
          description: Review invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewInvoiceResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /api/review-invoices/{invoiceId}/submit-tx:
    post:
      tags: [Reviews]
      summary: Submit a review payment transaction id
      operationId: submitReviewInvoiceTx
      parameters:
        - $ref: "#/components/parameters/InvoiceIdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReviewInvoiceTxRequest"
      responses:
        "200":
          description: Updated review invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewInvoiceResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "410":
          $ref: "#/components/responses/ErrorResponse"
  /api/trades:
    get:
      tags: [Trades]
      summary: List global trade history
      operationId: listTrades
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated global trade history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTradeHistoryResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /api/analytics/summary:
    get:
      tags: [Analytics]
      summary: Get site-wide access analytics summary
      operationId: getAnalyticsSummary
      parameters:
        - $ref: "#/components/parameters/HoursParam"
      responses:
        "200":
          description: Site-wide API access and token visit analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummaryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /api/analytics/endpoints:
    get:
      tags: [Analytics]
      summary: List analytics summaries for all tracked endpoints
      operationId: listEndpointAnalytics
      parameters:
        - $ref: "#/components/parameters/HoursParam"
      responses:
        "200":
          description: Per-endpoint access analytics summary rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndpointAnalyticsListResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /api/analytics/endpoints/{routeKey}:
    get:
      tags: [Analytics]
      summary: Get analytics detail for one endpoint
      operationId: getEndpointAnalytics
      parameters:
        - name: routeKey
          in: path
          required: true
          description: Fixed endpoint analytics route key
          schema:
            $ref: "#/components/schemas/AnalyticsRouteKey"
        - $ref: "#/components/parameters/HoursParam"
      responses:
        "200":
          description: Endpoint analytics detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndpointAnalyticsDetailResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /api/analytics/tokens:
    get:
      tags: [Analytics]
      summary: List token visit leaderboard rows
      operationId: listTokenVisits
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: sort
          in: query
          required: false
          description: Visit leaderboard sort field. Effective default is `visitsTotal`.
          schema:
            $ref: "#/components/schemas/TokenVisitSortField"
        - name: order
          in: query
          required: false
          description: Sort order. Effective default is `desc`.
          schema:
            $ref: "#/components/schemas/SortOrder"
      responses:
        "200":
          description: Paginated token visit leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTokenVisitSummaryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /api/analytics/tokens/{tokenId}:
    get:
      tags: [Analytics]
      summary: Get visit analytics for one token
      operationId: getTokenVisitAnalytics
      parameters:
        - $ref: "#/components/parameters/TokenIdPathParam"
        - $ref: "#/components/parameters/HoursParam"
      responses:
        "200":
          description: Token visit analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenVisitAnalyticsResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
components:
  parameters:
    TokenIdPathParam:
      name: tokenId
      in: path
      required: true
      description: Token identifier
      schema:
        type: string
    InvoiceIdPathParam:
      name: invoiceId
      in: path
      required: true
      description: Review invoice identifier
      schema:
        type: string
    PageParam:
      name: page
      in: query
      required: false
      description: 1-based page number. Default `1`.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      required: false
      description: Page size. Default `50`, max `200`.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    SortOrder:
      type: string
      enum: [asc, desc]
...
    TokenSortField:
      type: string
      enum:
        - totalTradeCount
        - totalVolumeSats
        - latestPriceNanosatsPerAtom
        - recent144TradeCount
        - recent144VolumeSats
        - recent1008TradeCount
        - recent1008VolumeSats
        - recent4320TradeCount
        - recent4320VolumeSats
        - lastTradeBlockHeight
        - lastTradeBlockTimestamp
...
    TokenDetailResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/TokenDetail"
...
    PaginatedTokenSummaryResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/PaginatedTokenSummary"
...
    PaginatedTradeHistoryResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/PaginatedTradeHistory"
...
    PaginatedTokenReviewsResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/PaginatedTokenReviews"
...
    TokenReviewSummaryResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/TokenReviewSummary"
...